requirements.txt
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
PyJWT>=2.8.0
python-multipart>=0.0.9
